.pi/cli/book-ingest/pyproject.toml
[project]
name = "book-ingest"
version = "0.3.0"
description = "Convert TTRPG PDFs into chunked, frontmattered Markdown via Marker."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "click>=8.1",
    "marker-pdf>=1.10.2,<2",
    "openai>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
book-ingest = "book_ingest.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/book_ingest"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

# --- Ruff: lint + format --------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "UP",   # pyupgrade
    "SIM",  # simplify
    "RUF",  # ruff-specific rules
    "C4",   # comprehensions
    "TID",  # tidy imports
]
ignore = [
    "E501",  # line too long — handled by formatter, not blocked
    "B008",  # function-call default (click decorators)
    "SIM108",  # ternary noise; sometimes the if/else reads better
    "RUF001",  # ambiguous en-dash; intentional in user-facing strings
    "RUF002",  # docstring en-dash; same
    "RUF003",  # comment en-dash; same
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011", "TID252"]

[tool.ruff.lint.isort]
known-first-party = ["book_ingest"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- Mypy: type checking --------------------------------------------------
[tool.mypy]
python_version = "3.11"
files = ["src/book_ingest", "tests"]
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_reexport = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = ["yaml", "click", "marker.*", "openai"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false

# --- Pytest --------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
